Refactoring/simplification of Repository tests, along with deletion of unused previous file.
It was a bit annoying to have one class test both repositories, so I split it in two. Still using a common inherited class to do most of the work.

ALSO, disabled parallelization for tests since having common inherited class somehow broke without it.
