+++ test/Chirp.CoreTest/AssemblyInfo.cs
@@ -1,0 +1,2 @@
+﻿// Made by ChatGPT
+[assembly: CollectionBehavior(DisableTestParallelization = true)]
